+++ public/scripts/plugins/pcr/views/home_pcr_view.js
@@ -90,6 +90,7 @@ export default Backbone.View.extend({
       displaySettings: {
         primaryView: primaryView
       },
+      tryShowingModal: true,
       sourceSequenceName: loadedSequence.name,
       features: loadedSequence.features,
     });